fix(audio): address review — update stale docstrings and rename MfccComputer to MelComputer
